Separate the JdkClient into an API to allow for other resources to be used like foojay.io.
